requirements.txt
flask
requests
pyjks
jsonschema
pyOpenSSL
paramiko